crud_service/requirements.txt
Django
djangorestframework
pymongo
python-decouple
